+++ uniffi_bindgen/src/macro_metadata/ci.rs
@@ -107,11 +107,7 @@ fn add_item_to_ci(iface: &mut ComponentInterface, item: Metadata) -> anyhow::Res
             iface.add_object_trait_impl(meta)?;
         }
         Metadata::CallbackInterface(meta) => {
-            iface.types.add_known_type(&Type::CallbackInterface {
-                module_path: meta.module_path.clone(),
-                name: meta.name.clone(),
-            })?;
-            iface.add_callback_interface_definition(CallbackInterface::try_from(meta)?);
+            iface.add_callback_interface_definition(CallbackInterface::try_from(meta)?)?;
         }
         Metadata::TraitMethod(meta) => {
             iface.add_trait_method_meta(meta)?;